Server/requirements.txt
autopep8==2.0.1
bcrypt==3.2.2
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10 
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
pandas